requirements.txt
fastapi[standard]>=0.111.0
pydantic>=2.7.0
sqlalchemy>=2.0.30
asyncpg>=0.29.0
redis>=5.0.0
pyjwt>=2.8.0
bcrypt>=4.0.0
uvicorn>=0.30.0
pytest
httpx
